Add support for updating the site-preferences.json file with updates read from the site model.
The idea is that we listen to any updates to the SitePreferences field
of the site model and write them into site-preferences.json file.

sphere-config includes this file. most applications just use sphere-config
(or the Go equivalent) to read the configuration.

The script client-helper.sh can be used to decide how to propagate the changes elsewhere in the client.

Signed-off-by: Jon Seymour <jon.seymour@gmail.com>