client-helper.sh: introduce a client helper script that is called by sphere-client when the preferences change.
Other daemons that might need to be restarted can provide their own
scripts that can apply a change in site preferences to their own state.
Such scripts should be located in /opt/ninjablocks/sphere-client/site-preference-helpers.

Note that it is the responsibility of the other scripts to
maintain and check a copy of the previous state in order to avoid
unnecessary restarts in case that relevant state has not changed.

Signed-off-by: Jon Seymour <jon@ninjablocks.com>